Fixed File order
Addressed the comment regarding ProjectOverview becoming dead, it will still be in use and usable here. Better organized the file structure of ProjectV2
